requirements.txt
commonmark==0.9.1
cowsay==4.0
loguru==0.5.3
rich==10.12.0
scapy==2.4.5
beautifulsoup4==4.10.0
bs4==0.0.1
soupsieve==2.3.1
python-nmap==0.7.1
pyreadline==2.1
scapy==2.4.5
mac-vendor-lookup==0.1.11
python_whois==0.7.3
requests==2.27.1